refactor: split CLI turn and approval logic
