+++ packages/agent-cli/test/Spec.hs
@@ -2,6 +2,7 @@ module Main (main) where
 
 import Test.Hspec (hspec)
 
+import qualified Agent.CLI.ApprovalSpec as ApprovalSpec
 import qualified Agent.CLI.AuthSpec as AuthSpec
 import qualified Agent.CLI.CancelWatchSpec as CancelWatchSpec
 import qualified Agent.CLI.ClipboardSpec as ClipboardSpec
@@ -30,6 +31,7 @@ import qualified Agent.CLI.WorktreeSpec as WorktreeSpec
 
 main :: IO ()
 main = hspec do
+    ApprovalSpec.spec
     AuthSpec.spec
     CancelWatchSpec.spec
     ClipboardSpec.spec
